include/starpu.h: fix typedef when using MSCV

+++ include/starpu.h
@@ -25,9 +25,9 @@
 #else
 typedef unsigned char uint8_t;
 typedef unsigned short uint16_t;
-typedef unsigned long uint32_t;
+typedef unsigned int uint32_t;
 typedef unsigned long long uint64_t;
-typedef unsigned long int uintptr_t;
+typedef unsigned int uintptr_t;
 #endif
 
 #include <starpu_config.h>
